+++ GloryOfEfficiency/Core/ModEntry.cs
@@ -102,17 +102,6 @@ namespace GloryOfEfficiency.Core
                 Logger.Log($"Don't Eat That(tm) is enabled!");
                 FoodAutomation.InitDontEat();
             }
-
-            Game1.graphics.PreferredBackBufferWidth = Conf.WindowWidth;
-            Game1.graphics.PreferredBackBufferHeight = Conf.WindowHeight;
-            var viewport = Game1.graphics.GraphicsDevice.Viewport;
-            viewport.Width = Conf.WindowWidth;
-            viewport.Height = Conf.WindowHeight;
-//            viewport.X = Conf.WindowX;             // Hackswell: seems to do nothing
-//            viewport.Y = Conf.WindowY;             // Hackswell: seems to do nothing
-//            viewport.Bounds = new Rectangle(new Point(Conf.WindowX, Conf.WindowY), new Point(Conf.WindowWidth, Conf.WindowHeight)); // Hackswell: seems to do nothing
-            GameRunner.instance.Window.Position = new Point(Conf.WindowX, Conf.WindowY);        // Hackswell: seems to do nothing
-            Game1.graphics.ApplyChanges();
         }
 
         private static void OnDebugCommand(string name, string[] args)